Python code generation/requirements.txt
streamlit
langchain
langchain-community
langchain-experimental
python-dotenv
openai